camunda/requirements.txt
coreapi==2.3.3
Django==3.1.1
django-configurations==2.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.3.0
django-solo==1.1.3
django-webtest==1.9.7
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-jwt-refresh-token==0.5
djangorestframework-simplejwt==4.4.0
docutils==0.16
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
